Contrast to Gtk.SpinButtons and Gdk version
The Gtk.SpinButtons in the toolbar had white background colour and white
colour for text inside them, making them not visible when active.
Also the version of Gdk was not specified before importing and there
was an unusual ordering of imports.

Now by changing their background colour to grey the text is more visible.
Now changing the order of imports and specifying version of Gdk before
import.